+++ src/Graph.php
@@ -22,16 +22,6 @@ use Innmind\Immutable\{
  */
 final class Graph
 {
-    /** @var T */
-    private string $directed;
-    private Name $name;
-    /** @var Set<Node> */
-    private Set $nodes;
-    /** @var Set<self<T>> */
-    private Set $clusters;
-    /** @var Map<string, string> */
-    private Map $attributes;
-
     /**
      * @param T $directed
      * @param Set<Node> $nodes
@@ -39,30 +29,22 @@ final class Graph
      * @param Map<string, string> $attributes
      */
     private function __construct(
-        string $directed,
-        Name $name,
-        Set $nodes,
-        Set $clusters,
-        Map $attributes,
-        Rankdir $rankdir = null,
+        private string $directed,
+        private Name $name,
+        private Set $nodes,
+        private Set $clusters,
+        private Map $attributes,
     ) {
-        $this->directed = $directed;
-        $this->name = $name;
-        $this->nodes = $nodes;
-        $this->clusters = $clusters;
-        $this->attributes = $attributes;
-
-        if ($rankdir) {
-            $this->attributes = ($this->attributes)('rankdir', $rankdir->toString());
-        }
     }
 
     /**
      * @psalm-pure
      *
+     * @param non-empty-string $name
+     *
      * @return self<'directed'>
      */
-    public static function directed(string $name = 'G', Rankdir $rankdir = null): self
+    public static function directed(string $name = 'G', ?Rankdir $rankdir = null): self
     {
         /** @var Set<Node> */
         $nodes = Set::of();
@@ -71,15 +53,21 @@ final class Graph
         /** @var Map<string, string> */
         $attributes = Map::of();
 
-        return new self('directed', Name::of($name), $nodes, $clusters, $attributes, $rankdir);
+        if ($rankdir) {
+            $attributes = ($attributes)('rankdir', $rankdir->toString());
+        }
+
+        return new self('directed', Name::of($name), $nodes, $clusters, $attributes);
     }
 
     /**
      * @psalm-pure
      *
+     * @param non-empty-string $name
+     *
      * @return self<'undirected'>
      */
-    public static function undirected(string $name = 'G', Rankdir $rankdir = null): self
+    public static function undirected(string $name = 'G', ?Rankdir $rankdir = null): self
     {
         /** @var Set<Node> */
         $nodes = Set::of();
@@ -88,7 +76,11 @@ final class Graph
         /** @var Map<string, string> */
         $attributes = Map::of();
 
-        return new self('undirected', Name::of($name), $nodes, $clusters, $attributes, $rankdir);
+        if ($rankdir) {
+            $attributes = ($attributes)('rankdir', $rankdir->toString());
+        }
+
+        return new self('undirected', Name::of($name), $nodes, $clusters, $attributes);
     }
 
     public function isDirected(): bool
@@ -140,6 +132,8 @@ final class Graph
     }
 
     /**
+     * @internal
+     *
      * @return Set<Node>
      */
     public function roots(): Set
@@ -156,6 +150,8 @@ final class Graph
     }
 
     /**
+     * @internal
+     *
      * @return Set<Node>
      */
     public function nodes(): Set
@@ -164,6 +160,9 @@ final class Graph
     }
 
     /**
+     *
+     * @param non-empty-string $label
+     *
      * @return self<T>
      */
     public function displayAs(string $label): self
@@ -228,6 +227,8 @@ final class Graph
     }
 
     /**
+     * @internal
+     *
      * @return Map<string, string>
      */
     public function attributes(): Map