+++ src/Graph/Name.php
@@ -11,27 +11,33 @@ use Innmind\Immutable\Str;
  */
 final class Name
 {
-    private string $value;
-
-    private function __construct(string $name)
-    {
-        if (!Str::of($name)->matches('~[a-zA-Z0-9_]+~')) {
-            throw new DomainException($name);
-        }
-
-        $this->value = $name;
+    /**
+     * @param non-empty-string $value
+     */
+    private function __construct(
+        private string $value,
+    ) {
     }
 
     /**
      * @psalm-pure
      *
+     * @param non-empty-string $name
+     *
      * @throws DomainException
      */
     public static function of(string $name): self
     {
+        if (!Str::of($name)->matches('~[a-zA-Z0-9_]+~')) {
+            throw new DomainException($name);
+        }
+
         return new self($name);
     }
 
+    /**
+     * @return non-empty-string
+     */
     public function toString(): string
     {
         return $this->value;