+++ src/Layout/DPI.php
@@ -3,34 +3,32 @@ declare(strict_types = 1);
 
 namespace Innmind\Graphviz\Layout;
 
-use Innmind\Graphviz\Exception\DomainException;
-
 /**
  * @psalm-immutable
  */
 final class DPI
 {
-    private int $value;
-
-    private function __construct(int $value)
-    {
-        if ($value < 1) {
-            throw new DomainException;
-        }
-
-        $this->value = $value;
+    /**
+     * @param int<1, max> $value
+     */
+    private function __construct(
+        private int $value,
+    ) {
     }
 
     /**
      * @psalm-pure
      *
-     * @throws DomainException
+     * @param int<1, max> $value
      */
     public static function of(int $value): self
     {
         return new self($value);
     }
 
+    /**
+     * @return int<1, max>
+     */
     public function toInt(): int
     {
         return $this->value;