+++ tests/Layout/DPITest.php
@@ -3,13 +3,10 @@ declare(strict_types = 1);
 
 namespace Tests\Innmind\Graphviz\Layout;
 
-use Innmind\Graphviz\{
-    Layout\DPI,
-    Exception\DomainException
-};
-use PHPUnit\Framework\TestCase;
+use Innmind\Graphviz\Layout\DPI;
 use Innmind\BlackBox\{
     PHPUnit\BlackBox,
+    PHPUnit\Framework\TestCase,
     Set,
 };
 
@@ -20,22 +17,9 @@ class DPITest extends TestCase
     public function testInterface()
     {
         $this
-            ->forAll(Set\Integers::above(1))
+            ->forAll(Set::integers()->above(1))
             ->then(function(int $int): void {
                 $this->assertSame($int, DPI::of($int)->toInt());
-            });
-    }
-
-    public function testThrowWhenLowerThanOne()
-    {
-        $this
-            ->forAll(Set\Integers::below(0))
-            ->then(function(int $int): void {
-                try {
-                    DPI::of($int);
-                } catch (\Throwable $e) {
-                    $this->assertInstanceOf(DomainException::class, $e);
-                }
             });
     }
 }