+++ tests/Node/NameTest.php
@@ -7,9 +7,9 @@ use Innmind\Graphviz\{
     Node\Name,
     Exception\DomainException
 };
-use PHPUnit\Framework\TestCase;
 use Innmind\BlackBox\{
     PHPUnit\BlackBox,
+    PHPUnit\Framework\TestCase,
     Set,
 };
 
@@ -20,20 +20,13 @@ class NameTest extends TestCase
     public function testInterface()
     {
         $this
-            ->forAll(Set\Strings::any())
+            ->forAll(Set::strings())
             ->filter(static function(string $string): bool {
                 return \strlen($string) > 0 && \strpos($string, '->') === false && \strpos($string, '-') === false && \strpos($string, '.') === false && \strpos($string, "\x00") === false;
             })
             ->then(function(string $string): void {
                 $this->assertSame($string, Name::of($string)->toString());
             });
-    }
-
-    public function testThrowWhenEmpty()
-    {
-        $this->expectException(DomainException::class);
-
-        Name::of('');
     }
 
     public function testThrowWhenContainingAnArrow()