+++ CHANGELOG.md
@@ -1,5 +1,33 @@
 # Changelog
 
+## 4.0.0 - 2025-04-18
+
+### Added
+
+- Support for `innmind/filesystem:~8.0`
+
+### Changed
+
+- Use more precise primitive types
+- `Innmind\Graphviz\Attribute\Value` is now internal
+- `Innmind\Graphviz\Node::attributes()` is now internal
+- `Innmind\Graphviz\Node\Shape::attributes()` is now internal
+- `Innmind\Graphviz\Graph::roots()` is now internal
+- `Innmind\Graphviz\Graph::nodes()` is now internal
+- `Innmind\Graphviz\Graph::attributes()` is now internal
+- `Innmind\Graphviz\Edge::attributes()` is now internal
+
+### Removed
+
+- Support for `innmind/immutable:~4.0`
+- `Innmind\Graphviz\Edge\Shape::open()`
+- `Innmind\Graphviz\Edge\Shape::left()`
+- `Innmind\Graphviz\Edge\Shape::right()`
+
+### Fixed
+
+- PHP `8.4` deprecations
+
 ## 3.4.0 - 2023-10-22
 
 ### Changed